+++ src/java/dao/AuditDAO.java
@@ -4,5 +4,5 @@ package dao;
  * Interface for Audit Log Database Operations
  */
 public interface AuditDAO {
-    void log(Long userId, String action, String entity, Long entityId);
+    void log(Integer userId, String action, String entity, Integer entityId);
 }
